pyproject.toml
[tool.poetry]
name = "prometheus_block_exporter"
version = "1.0.1"
description = "A tool for exporting Prometheus TSDB Blocks."
authors = ["Lajos Harangi <lajos@harangi.dev>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.11"
python-ulid = "^1.1.0"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
flake8-import-order = "^0.18.1"
pep8-naming = "^0.13.2"
ipython = "^8.6.0"
flake8 = "^4.0.1"
flakeheaven = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 300

[tool.flakeheaven]
max_line_length = 300
exclude = ["__pycache__", ".venv", "docs"]

[tool.flakeheaven.plugins]
"*" = ["+*"]
